pick chars from password by char position
Several banking websites in the UK require you to extract certain
characters from your password when logging in or purchasing something.
This is annoying when all your passwords are generated, so I implemented
an input field next to each entry in which you can enter the char
positions delimited by spaces (e.g. 1 8 15). Clicking the input field
then sets the value of the input field to the characters at those
positions in the password string (e.g. eX*)

Didn't spend a lot of time making it look nice - works for me